fixed a lot of bugs

+++ server/Trivia_Server/Trivia_Server/JsonResponsePacketSerializer.h
@@ -65,7 +65,7 @@ struct GetRoomStateResponse {
 	bool hasGameBegun;
 	std::vector<std::string> players;
 	unsigned int questionCount;
-	float answerTimeOut;
+	unsigned int answerTimeOut;
 };
 
 struct LeaveRoomResponse {
